+++ backend/src/main/java/com/dando_la_vez/backend/services/PuestoService.java
@@ -19,7 +19,7 @@ public class PuestoService {
         return puestoRepository.findAll();
     }
 
-    public Optional<Puesto> getPuestoId(long id){
+    public Optional<Puesto> getPuestoId(int id){
         return puestoRepository.findById(id);
     }
 
@@ -31,7 +31,7 @@ public class PuestoService {
         return puestoRepository.save(puesto);
     }
 
-    public void deletePuesto(long id){
+    public void deletePuesto(int id){
         puestoRepository.deleteById(id);
     }
